requirements-web.txt
Flask>=2.3.0
Werkzeug>=2.3.0
gunicorn>=21.0.0
selenium>=4.15.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-image>=0.21.0
scikit-learn>=1.3.0
webdriver-manager>=4.0.0
matplotlib>=3.7.0
reportlab>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
